Order repo controller

+++ OfferOasisBackend/OfferOasisBackend/Model/Order.cs
@@ -9,7 +9,7 @@ public class Order
     private string BillingAddress { get; init; }
     private string ShippingAddress { get; init; }
     private string ShippingType { get; init; } // Todo: lehetne enum
-    private double ShippingCost { get; init; }
+    private decimal ShippingCost { get; init; }
     private DateTime OrderDate { get; init; }
     private User _user;
 }
